Resolves two conflicts against main b4ae3eb0:

- larql-server bootstrap/load.rs: keep this branch's doc comment on
  unsupported_v3_options — the branch implements V3 layer-prefix
  sharding, so main's "fail closed until V3 sharding exists" no longer
  describes the function.
- larql-vindex opplan/exec/decode.rs: keep both new Entry variants.
  Hidden (step_embedding) is an external embedding row that enters the
  declared topology as an embedding would; Single (main's
  step_from_carrier_intervened) is a single-stream carrier entering a
  layer-range slice. Different contracts, so neither replaces the other.

+++ crates/larql-compute-metal/src/shaders/q4k_grouped_experts.rs
@@ -44,7 +44,7 @@ kernel void q4k_grouped_experts(
 
     const uint superblocks   = K / 256u;
     const uint bytes_per_row = superblocks * Q4KGE_BLOCK_SIZE;
-    device const uchar* row_w = W4K + offsets[slot] + row_idx * bytes_per_row;
+    device const uchar* row_w = W4K + (ulong)offsets[slot] + (ulong)row_idx * bytes_per_row;
     device const float* Xs = X + (ulong)slot * XSTRIDE;
 
     // 2-way inter-superblock interleaving.
